Set properties that potentially "late init" to mutable
The Cross Platform mobile SDKs have certain telemetry properties that cannot be known until after the call for initialization. Before we send the telemetry configuration, we need these SDKs to be able to modify certain properties before we send them to telemetry.
Also removed a Flutter specific property 'attachToExisting' that cannot be known early enough in the app lifecycle to be sent with telemetry.