👌 modify entrypoints consumed by the iOS SDK
To simplify things a bit, the iOS SDK will use
`schemas/rum-events-mobile-schema.json` and
`schemas/session-replay-mobile-format.json` as entrypoints instead of
`rum-events-format.json` and `session-replay-mobile-format.json`.

To accommodate this change, this commit includes the telemetry events in
the rum mobile entry point.

This commit also clarifies how all this is used in the DataDog
ecosystem.